Update README.md
One customer (as well as myself) did not immediately realize that this script is only for plugins subscribers. They wrote to us and asked "How come we are running this and it is asking for a key. What key is that?" (Customer has no plugins subscriptions and lost 8 days to our back and forth on this issue.)  

My edit adds clarity: "this script is for plugins subscribers".